Initial generation code and implementation data
The generation code was split out of the web-roadmaps project:
https://github.com/w3c/web-roadmaps

The code is basically the same. It has just been re-written in a more
modularized way. Basic tests have been added too.

The structure of data files in the `data` folder has been slightly adjusted
compared to the web-roadmaps project. The initial data files in the `data`
are the result of applying a small conversion tool. Data files in web-roadmaps
that actually describe a specific spec level or a specific feature within a
specification have not been integrated yet (~60 features missing for now).

The `index.json` generation logic runs as a GitHub action once per day. It was
inspired from the browser-specs project:
https://github.com/w3c/browser-specs