Fully separate info for Safari and Webkit
The webkit platform status site reports implementation about Webkit. There is
no guarantee that a feature that is supported in Webkit is also supported in
Safari or Safari on iOS. The code just takes for granted that when the support
status of a feature reported by the Webkit platform status site is lower than
the status reported by some other project, then the info from the Webkit
platform status site should be assumed to be correct.

The code had several flaws though:
1. "Partially supported" in the Webkit platform status data was considered to
mean supported. Partial flag is now set when that happens.
2. Support information for Safari and Webkit were not fully separated, meaning
that when the info from the Webkit platform status site was preferred, there
was no `selected` info for `safari` or `safari_ios`.
3. The code that preferred the info from the Webkit platform status site was
actually not working...
4. In general, there would be no selected info for the `webkit` UA.

This update fixes all bugs, and fully separates the `safari` and `safari_ios`
info from the `webkit` info. When the info about `webkit` needs to be selected
for `safari` or `safari_ios`, a new support entry gets created in the support
array.

A new test was added to make sure that, for each UA featured in a support array,
there exists a `selected` entry.