Major overhaul: Automate mappings with status platforms (#10)
This major update finds and proposes new possible mappings between specs and
entry points in implementation status platforms, matching on spec URLs.

Breaking changes:
- The structure of data files changes to always take objects to describe a
mapping and to allow more than one mapping for a spec. This makes the data file
slightly less convenient to edit manually, but then the intent is for it to be
mainly managed automatically from now on.
- The structure of index files changes slightly to replace the `href` property
with a `details` property that lists all individual support information.

README still needs to be updated and job automation still needs to be done!