Refresh existing mappings
This drops mappings that are no longer accurate, validates mappings that are
still good but that cannot be automatically associated with the spec at hand,
and lists a number of alternative URLs that are used to reference specs.

With this update, the find-mappings report currently no longer contains any
request to drop a mapping.