Add features referenced in web-roadmaps
This adds a series of spec features that web roadmaps talk about. This also
adds missing "manual" entries for some of the specs and the possibility to
have an empty string status to mean "we don't know but we know that other status
values returned by status platforms are wrong".