Implementation status: don't guess based on a guess
Guessing an implementation status is by definition unreliable. Guessing an
implementation status based on a status that was already a guess all the more
so.

This update skips guesses when trying to look for better implementation status
info (for webkit) or when trying to guess the implementation status at a
feature level from spec support.