Fix implementation status algorithm
This fixes a few bugs related to the `partial`, `representative` and `selected`
flag in the final implementation status info, in particular due to the fact that
the code mixed partial and representative under one concept, whereas it should
be treated seperately.

The `representative` flag is now reported in the `index.json` file.

The update also includes a few data fixes or improvements.